refactor(logstream): move LogStream header to detail
Relocate LogStream.hpp into the detail directory and fix dependent includes.
